backend-example-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
pyjwt==2.8.0

# Banco de dados (escolha um)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.1          # MongoDB
# sqlalchemy==2.0.25      # ORM genérico
